reorganize hodge files
